[refactor] Move Slice class to array module.
[python/assassyn/ir/array.py] Add Slice class.
[python/assassyn/ir/expr/expr.py] Remove Slice class, import Slice in is_valued method.
[python/assassyn/ir/value.py] Update import to use Slice from array module.
[python/assassyn/codegen/verilog/design.py] Update import to use Slice from array module.
[python/assassyn/codegen/simulator/modules.py] Update import to use Slice from array module.
