+++ python/assassyn/ir/array.py
@@ -16,6 +16,50 @@ if typing.TYPE_CHECKING:
     from .module.base import ModuleBase
 
 
+class Slice(Expr):
+    '''The class for slice operation, where x[l:r] as a right value'''
+
+    SLICE = 700
+
+    def __init__(self, x, l: int, r: int):
+        assert isinstance(l, int), f'Only int literal can slice, but got {type(l)}'
+        assert isinstance(r, int), f'Only int literal can slice, but got {type(r)}'
+        assert isinstance(x, Value), f'{type(x)} is not a Value!'
+        l = to_uint(l)
+        r = to_uint(r)
+        super().__init__(Slice.SLICE, [x, l, r])
+
+    @property
+    def x(self) -> Value:
+        '''Get the value to slice'''
+        return self._operands[0]
+
+    @property
+    def l(self) -> int:
+        '''Get the value to slice'''
+        return self._operands[1]
+
+    @property
+    def r(self) -> int:
+        '''Get the value to slice'''
+        return self._operands[2]
+
+    @property
+    def dtype(self) -> DType:
+        '''Get the data type of the sliced value'''
+        # pylint: disable=import-outside-toplevel
+        from .dtype import Bits
+        from .const import Const
+        assert isinstance(self.l.value, Const)
+        assert isinstance(self.r.value, Const)
+        return Bits(self.r.value.value - self.l.value.value + 1)
+
+    def __repr__(self):
+        l = self.l.as_operand()
+        r = self.r.as_operand()
+        return f'{self.as_operand()} = {self.x.as_operand()}[{l}:{r}]'
+
+
 
 def RegArray( #pylint: disable=invalid-name,too-many-arguments
         scalar_ty: DType,
